# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.CreateOrderRequest:
    properties:
      client_id:
        type: integer
      total:
        type: number
    required:
    - client_id
    - total
    type: object
...
  controllers.UpdateOrderPaymentRequest:
    properties:
      amount:
        example: 99.99
        type: number
      payment_date:
        example: "2024-11-03T00:00:00Z"
        type: string
      payment_desc:
        example: Credit card payment
        type: string
      payment_method:
        example: 1
        type: integer
      payment_ref:
        example: REF123456
        type: string
      payment_status:
        example: paid
        type: string
    required:
    - amount
    - payment_status
    type: object
...
host: test.hamber-hub.com
info:
  contact:
    email: support@hamber-hub.com
    name: API Support
    url: http://www.hamber-hub.com/support
  description: This is the API documentation for the Hamber platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Hamber API Documentation
  version: "1.0"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: Get paginated list of user orders
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Orders list
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get orders list
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create a new order for products
      parameters:
      - description: Order details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Order created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create a new order
      tags:
      - Orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order details
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get order by ID
      tags:
      - Orders
  /orders/{id}/cancel:
    patch:
      consumes:
      - application/json
      description: Cancel an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order cancelled
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid order ID
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Cancel order
      tags:
      - Orders
  /orders/{id}/payment:
    put:
      consumes:
      - application/json
      description: Updates the payment status, amount, reference, date, method, and
        description for a specific order.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payment update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateOrderPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Order payment updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid order ID or bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update order payment
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update order payment details
      tags:
      - Orders
  /orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Status updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update order status
      tags:
      - Orders
...
swagger: "2.0"